+++ packages/world/tasks/primitives/wander-task/__tests__/execute.spec.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it, jest } from '@jest/globals';
 import { defaultActorConstructorOptions, MockedActor } from '@sentient-world/engine/mocks';
 
-import type { DateTime } from '../../../../core';
-import { MockedComparator } from '../../../../core/mocks';
+import type { DateTime } from '../../../../date-time';
+import { MockedComparator } from '../../../../date-time/mocks';
 import { mockedState, StateBuilder } from '../../../../state/mocks';
 
 import { getMockedWanderTaskDeps } from '../mocks';
